Update type hints for compatibility with NumPy 2.x

+++ tests/cluster_recipes/test_murata_binned_spec_z.py
@@ -69,8 +69,8 @@ def test_get_theory_prediction_returns_value(
     assert prediction is not None
     assert callable(prediction)
 
-    mass = np.linspace(13, 17, 2)
-    z = np.linspace(0.1, 1, 2)
+    mass = np.linspace(13, 17, 2, dtype=np.float64)
+    z = np.linspace(0.1, 1, 2, dtype=np.float64)
     mass_proxy_limits = (0, 5)
     sky_area = 360**2
 
@@ -85,8 +85,8 @@ def test_get_theory_prediction_with_average_returns_value(
     cluster_abundance: ClusterAbundance,
     murata_binned_spec_z: MurataBinnedSpecZRecipe,
 ):
-    mass = np.linspace(13, 17, 2)
-    z = np.linspace(0.1, 1, 2)
+    mass = np.linspace(13, 17, 2, dtype=np.float64)
+    z = np.linspace(0.1, 1, 2, dtype=np.float64)
     mass_proxy_limits = (0, 5)
     sky_area = 360**2
 
@@ -141,8 +141,8 @@ def test_get_theory_prediction_throws_with_nonimpl_average(
     assert prediction is not None
     assert callable(prediction)
 
-    mass = np.linspace(13, 17, 2)
-    z = np.linspace(0.1, 1, 2)
+    mass = np.linspace(13, 17, 2, dtype=np.float64)
+    z = np.linspace(0.1, 1, 2, dtype=np.float64)
     mass_proxy_limits = (0, 5)
     sky_area = 360**2
